+++ frontend/src/components/Notification/useNotification.ts
@@ -1,0 +1,6 @@
+import React from 'react'
+import { NotificationContext } from './NotificationProvider'
+
+const useNotification = () => React.useContext(NotificationContext)
+
+export default useNotification
